+++ packages/admin/resources/lang/en/orders.php
@@ -1,0 +1,21 @@
+<?php
+
+return [
+    /**
+     * Order listing.
+     */
+    'index.title' => 'Orders',
+    'index.export_btn' => 'Export Orders',
+    'index.save_search.title' => 'Save Search',
+    'index.save_search.btn' => 'Save Search',
+    'index.all_orders' => 'All Orders',
+    'index.search_placeholder' => 'Search...',
+    'index.update_status.btn' => 'Update Status',
+    'index.update_status.title' => 'Update Status',
+    'index.delete_saved_search.title' => 'Delete saved search',
+    'index.delete_saved_search.confirm' => 'Are you sure you want to delete this saved search?',
+    'index.delete_saved_search.btn' => 'Confirm Delete',
+    'update_status.btn' => 'Update Status',
+    'update_status.title' => 'Update Status',
+    'show.title' => 'Order #:id',
+];