+++ packages/admin/resources/views/components/orders/status.blade.php
@@ -1,0 +1,3 @@
+<span class="inline-block px-2 py-1 text-xs text-white rounded" style="background: {{ $color }};">
+  {{ $label }}
+</span>