+++ packages/core/database/factories/OrderAddressFactory.php
@@ -20,7 +20,7 @@ class OrderAddressFactory extends Factory
             'line_two'              => $this->faker->boolean ? $this->faker->secondaryAddress : null,
             'line_three'            => $this->faker->boolean ? $this->faker->buildingNumber : null,
             'city'                  => $this->faker->city,
-            'state'                 => $this->faker->boolean ? $this->faker->county : null,
+            'state'                 => $this->faker->boolean ? $this->faker->state : null,
             'postcode'              => $this->faker->postcode,
             'delivery_instructions' => $this->faker->boolean ? $this->faker->sentence : null,
             'contact_email'         => $this->faker->boolean ? $this->faker->safeEmail : null,