+++ packages/core/src/GetCandyServiceProvider.php
@@ -36,6 +36,7 @@ use GetCandy\Models\Channel;
 use GetCandy\Models\Collection;
 use GetCandy\Models\Currency;
 use GetCandy\Models\Language;
+use GetCandy\Models\Order;
 use GetCandy\Models\OrderLine;
 use GetCandy\Models\Url;
 use GetCandy\Observers\AddressObserver;
@@ -45,6 +46,7 @@ use GetCandy\Observers\CollectionObserver;
 use GetCandy\Observers\CurrencyObserver;
 use GetCandy\Observers\LanguageObserver;
 use GetCandy\Observers\OrderLineObserver;
+use GetCandy\Observers\OrderObserver;
 use GetCandy\Observers\UrlObserver;
 use Illuminate\Auth\Events\Login;
 use Illuminate\Auth\Events\Logout;
@@ -227,6 +229,7 @@ class GetCandyServiceProvider extends ServiceProvider
         Url::observe(UrlObserver::class);
         Collection::observe(CollectionObserver::class);
         CartLine::observe(CartLineObserver::class);
+        Order::observe(OrderObserver::class);
         OrderLine::observe(OrderLineObserver::class);
         Address::observe(AddressObserver::class);
     }